Remove a lot of hardcoding, began leveraging dataset-api-tools
Now leveraging the dataset-api-tools library to standardize dataset/schema structure. Had to backport the lib to java 8, which this relies on. THIS COMMIT IS NOT PRODUCTION READY. The version on that lib is incorrect (my local machine only). /dataset/{id} not responds with the dataset, but in the wrong format. /datasets is likely broken. 

ALL TESTS HAVE BEEN COMMENTED OUT OR ARE OUTRIGHT BROKEN, BUILD BY SKIPPING TESTS.